### 2021-08-27 #3665 @whedon
[PRE REVIEW]: cppTPSA/pyTPSA: a C++/Python package for truncated power series algebra
**Submitting author:** @zhanghe9704 (<a href="http://orcid.org/0000-0001-7701-4118">He Zhang</a>)
**Repository:** https://github.com/zhanghe9704/tpsa
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
