+++ paper/paper.bib
@@ -70,12 +70,14 @@ publisher   = "Academic Press",
 address     = "San Diego",
 year        = "1999",
 ISBN        = "0-12-014750-5",
+doi			= "10.1016/s1076-5670(08)x7018-1",
 note        = "{A}lso available at http://bt.pa.msu.edu/pub"}
 
 @techreport{chao2002lecture,
   title={Lecture notes on topics in accelerator physics},
   author={Chao, A. W.},
   year={2002},
+  doi={10.2172/812598},
   institution={Stanford Linear Accelerator Center, Menlo Park, CA (US)}
 }
 
@@ -85,7 +87,8 @@ title       = "{COSY INFINITY} Version 9",
 journal     = "Nuclear Instruments and Methods",
 volume      = "558",
 pages       = "346-350",
-year        = "2005"}
+doi 		= "10.1016/j.nima.2005.11.109",
+year        = "2006"}
 
 @article{forest2002introduction,
   title={Introduction to the polymorphic tracking code},
@@ -108,6 +111,7 @@ year        = "2005"}
   author={Massari, M. and Di Lizia, Pierluigi and Cavenago, Francesco and Wittig, Alexander},
   booktitle={2018 AIAA Information Systems-AIAA Infotech@ Aerospace},
   pages={0398},
+  doi={10.2514/6.2018-0398},
   year={2018}
 }
 
