### 2021-08-31 576d21a
Add a few more DOIs.
