+++ paper/paper.bib
@@ -14,6 +14,7 @@ title       = "High-Order Computation and Normal Form Analysis of
                Physics of Particle Accelerators",
 volume      = "249",
 pages       = "456",
+doi			= "10.1063/1.41975",
 publisher   = "American Institute of Physics",
 address     = "New York",
 year        = "1991"}
@@ -23,6 +24,7 @@ author      = "M. Berz and K. Makino",
 title       = "Verified Integration of {ODE}s and Flows using
                Differential Algebraic Methods on High-Order {T}aylor Models",
 journal     = "Reliable Computing",
+doi			= "10.1023/A:1024467732637",
 volume      = "4",
 number      = "4",
 pages       = "361-369",
@@ -41,7 +43,7 @@ title = "The Fast Multipole Method in the Differential Algebra Framework",
 journal = " Nuclear Instruments and Methods A 645",
 pages = {338-344},
 year = "2011",
-note = "",
+doi = "10.1016/j.nima.2011.01.053",
 author = "H. Zhang and M. Berz"
 }
 
