### 2021-09-10 #3665 @zhanghe9704
Hello Dr. Smith,

Indeed I was hesitant about whether to include Lingyun as an author. The
reason that I didn't are (1) Lingyun's code was included in the previous
versions of MAD-X so it has been published somewhere else as part of the
MAD-X; and (2) except for the two files used in the cppTPSA repository He
has not been involved in the work of the two repositories of the paper,
although I did contact him and get his permission before I put the codes on
GitHub.

If you think it's better to include him as an author, I can contact him.

Thanks and best regards,

He Zhang
Computational Accelerator Physicist
Center for Advanced Science of Accelerators
Jefferson Lab
(O) 757-269-6657
www.hezhang.us

On Fri, Sep 10, 2021 at 4:00 AM Arfon Smith ***@***.***>
wrote:

> Thanks to Dr. Lingyun Yang for providing his tpsa code.
>
> @zhanghe9704 <https://github.com/zhanghe9704> - I noticed this comment in
> your README and wondered if you had considered including them as an author
> of your submission?
>
