### 2021-10-12 #3665 @sjcrs
I'd like to help, but I think I have a COI - my company works with He Zhang's JSPEC code, and have publications in progress that may involve him as an author (He, I'll be sending you an email about that soon).
