### 2022-03-05 #3665 @zhanghe9704
Here is a paper on the tpsa package used in MAD-X:

https://cds.cern.ch/record/2141771/files/mopje039.pdf

He
