### 2022-05-03 #3665 @zhanghe9704
Hello  @pibion,

Thanks for your reply. Have you contacted Prof. Yue Hao ***@***.***)
at Michigan State University? He may be able to help find a reviewer.

Dr. Ji Qiang ***@***.***) at Lawrence Berkeley National Lab may also be
able to help.

Another idea is to contact the developers or users of the DACE library:
https://github.com/dacelib/dace. I  think they should know about TA/TPSA.
COI: You may find Alex Wittig as the main developer of DACE. We graduated
from the same group with the same Ph.D. adviser. We actually sit in the
same office for about six years.

Best regards,
He Zhang
Computational Accelerator Physicist
Center for Advanced Science of Accelerators
Jefferson Lab
(O) 757-269-6657
www.hezhang.us
