### 2022-05-09 #3665 @pibion
@dtabell, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.

Please let me know if you have any questions, I'm happy to communicate in whatever way works best for you - just let me know!
