### 2022-08-05 #3665 @arfon
@pibion – it looks like you might need to find new reviewers here. This might be a good source of candidates: https://github.com/topics/power-series
