### 2022-10-18 #4818 @chrisrogers1234
@zhanghe9704 @pibion I attach referees comments.

### Summary
The code seems to be interesting. However, it is very hard to penetrate what is
doing what to which code block without some further documentation or overview. I
was able to build and run the example but not the tests.

[2022-10-17_reviewer-comments.txt](https://github.com/openjournals/joss-reviews/files/9811088/2022-10-17_reviewer-comments.txt)
