### 2022-10-18 #4818 @zhanghe9704
@chrisrogers1234 <https://github.com/chrisrogers1234>
Sorry for the delay. It takes some time for me to remember what I did since
I haven't touched the code for about a year.

Turns out the Makefile for the tests is not included in the online
repository. I attached it here.

You need to use cmake and make to compile the libs first.
Then in the test folder, use make with Makefile to compile the tests. (The
libs from the above step will be needed and will be found automatically.)

An alternative way is to compile the libs using the cbp project and then
compile the test using the cbp project if you use CodeBlocks.
I actually use CodeBlocks to develop the code. The cmake file and Makefiles
were added and tested after the cbp projects.

Thanks for reviewing the submission.

Best regards,
He

On Mon, Oct 17, 2022 at 4:59 AM chrisrogers1234 ***@***.***>
wrote:

> @zhanghe9704 <https://github.com/zhanghe9704>
>
> Could you advise how to build the tests? I tried doing something like gcc
> tests.cc -o test -L../ -ltpsa -lstdc++ -lm -std=c++14 and get loads of
> "undefined references" to Catch::. I note Catch is an external test package
> but it looks like you have copied some source code into your library.
> Should I build Catch as an external dependency or is it supposed to be
> built as part of the tests (and if so what is the recipe?). Can I build the
> tests from within cmake/make or is it like the examples?
>
> Many thanks!
>
