### 2022-10-18 #4818 @chrisrogers1234
@zhanghe9704

Thanks! I will have a look, maybe later this week. Also I attached a few recommendations above. The code looks good, but I think I need a bit more documentation just to orient my brain to help with the review.
