### 2022-10-27 #4818 @pibion
@zhanghe9704 would you be willing to add the makefile for the test compilation to the code repository?  I don't think it attached (or perhaps I'm not seeing it) in this issue thread.  And having it in the repository itself would be helpful.
