### 2022-11-03 #4818 @pibion
@zhanghe9704 I'm having trouble compiling the libraries and binaries.  I've tried with CodeBlocks, but it seems to be having trouble recognizing that the default compiler is gcc.  I've tried opening and building `cbp/tpsa_dll.cbp` and `cbp/tpsa_lib.cbp`.  In both cases I get the following error:

```
The compiler's setup is invalid, so Code::Blocks cannot find/run the compiler.
  Probably the toolchain path within the compiler options is not setup correctly?!
  Do you have a compiler installed?
Goto "Settings->Compiler...->Global compiler settings->unknown->Toolchain executables" and fix the compiler's setup
```

I suspect the issue here is that I don't actually have a gcc compiler installed that codeblocks can find, so I'll work on that.

In the meantime, I tried building the code in WSL, where I have a new-ish version of gcc installed (9.4.0).  When I first ran `cmake .`, I got an error saying CMake couldn't find the C++ compiler.  Per its suggestion I set an environment variable, `CXX`, pointing to my gcc.  But then I get an error saying CMake couldn't compile a test program.

Are there any other environmental variables you set?
