### 2022-11-03 #4818 @pibion
@zhanghe9704 I was able to compile the libraries and tests in WSL by installing g++ :D.  However, when I try to run the `tests` executable, I get a segmentation fault in addition to the tests passing:

```
(base) aroberts@LAPTOP-74V3RK8A:/mnt/c/Users/canto/Repositories/tpsa/test$ ./tests
===============================================================================
All tests passed (21 assertions in 3 test cases)

Segmentation fault
```

Have you encountered this before?

Still working on Codeblocks, if you have any suggestions I'd be grateful!
