### 2022-12-16 #4818 @chrisrogers1234
@danielskatz

Did you manage to address the documentation requests in this note?

https://github.com/openjournals/joss-reviews/files/9811088/2022-10-17_reviewer-comments.txt
