### 2022-12-16 #4818 @danielskatz
@chrisrogers1234 - I'm the track editor here, so I don't think you meant this for me, but for the author, @zhanghe9704

As a side note, this would be easier to see and work with if it was not an attachment, and was instead a bunch of issues in the source respository, each including a tag to this review (ie a mention of `openjournals/joss-reviews#4818`) - this will create a link to this thread, and here, we'll be able see if each is still open or closed.
