### 2022-12-16 #4818 @zhanghe9704
@chrisrogers1234 - Sorry, I haven't worked on that. Unfortunately, I'm very
busy now with two proposals due in early January.
I don't think I'll have much time to work on this till the end of the year
but I'll try.
BTW, I'm new to this process and appreciate any instructions.

Thanks for all your help. Happy holiday!

On Fri, Dec 16, 2022 at 7:35 AM chrisrogers1234 ***@***.***>
wrote:

> @danielskatz <https://github.com/danielskatz>
>
> Did you manage to address the documentation requests in this note?
>
>
> https://github.com/openjournals/joss-reviews/files/9811088/2022-10-17_reviewer-comments.txt
>
