### 2022-12-16 #4818 @danielskatz
Ideally, the process here is that reviewers find problems that keep them from checking off all the items on their checklist, and express those problems either as comments here (if very simple) or as issue or PRs in your repository. You try to work on them, or discuss why you think you don't need to. The editor (@pibion) moderates the discussion as needed to get to consensus, and makes decisions if there doesn't seem to be consensus.  This process continues until the submission is acceptable, or until you decide to stop it, and the whole process ideally takes 4-6 weeks, or a few months at most.
