+++ doc/doxygen/html/search/all_3.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['c_2b_2b_20tpsa_20lib_50',['C++ TPSA Lib',['../md__e___work__code_tpsa_lib_readme.html',1,'']]],
-  ['clean_51',['clean',['../struct_d_a_vector.html#ad4d594d154f7e8ec5006e3e223e4cdea',1,'DAVector::clean(const double eps)'],['../struct_d_a_vector.html#aa1494c69478b3b57d5973e86248a7e63',1,'DAVector::clean()']]],
-  ['compare_5fcd_5fvectors_52',['compare_cd_vectors',['../da_8cc.html#acc57aeb4ed529889b0eb2eb856599eaf',1,'da.cc']]],
-  ['compare_5fcd_5fwith_5ffile_53',['compare_cd_with_file',['../da_8cc.html#a1bf2bf9ee390a96ae6b0b2db05ca1fe1',1,'da.cc']]],
-  ['compare_5fda_5fvectors_54',['compare_da_vectors',['../da_8h.html#a9f08d1308badc3e175572d7c25948604',1,'compare_da_vectors(DAVector &amp;a, DAVector &amp;b, double eps=1e-15):&#160;da.cc'],['../da_8cc.html#a901411b03028f0c3a763b637e28cddd5',1,'compare_da_vectors(DAVector &amp;a, DAVector &amp;b, double eps):&#160;da.cc']]],
-  ['compare_5fda_5fwith_5ffile_55',['compare_da_with_file',['../da_8cc.html#aa0d7d543f79a6281a02af7e318ca6d65',1,'da.cc']]],
-  ['con_56',['con',['../struct_d_a_vector.html#a00ef34ddb1dce8bead6d43a24b3fb23e',1,'DAVector']]]
+  ['clean_50',['clean',['../struct_d_a_vector.html#ad4d594d154f7e8ec5006e3e223e4cdea',1,'DAVector::clean(const double eps)'],['../struct_d_a_vector.html#aa1494c69478b3b57d5973e86248a7e63',1,'DAVector::clean()']]],
+  ['compare_5fcd_5fvectors_51',['compare_cd_vectors',['../da_8cc.html#acc57aeb4ed529889b0eb2eb856599eaf',1,'da.cc']]],
+  ['compare_5fcd_5fwith_5ffile_52',['compare_cd_with_file',['../da_8cc.html#a1bf2bf9ee390a96ae6b0b2db05ca1fe1',1,'da.cc']]],
+  ['compare_5fda_5fvectors_53',['compare_da_vectors',['../da_8h.html#a9f08d1308badc3e175572d7c25948604',1,'compare_da_vectors(DAVector &amp;a, DAVector &amp;b, double eps=1e-15):&#160;da.cc'],['../da_8cc.html#a901411b03028f0c3a763b637e28cddd5',1,'compare_da_vectors(DAVector &amp;a, DAVector &amp;b, double eps):&#160;da.cc']]],
+  ['compare_5fda_5fwith_5ffile_54',['compare_da_with_file',['../da_8cc.html#aa0d7d543f79a6281a02af7e318ca6d65',1,'da.cc']]],
+  ['con_55',['con',['../struct_d_a_vector.html#a00ef34ddb1dce8bead6d43a24b3fb23e',1,'DAVector']]],
+  ['cpptpsa_20_2d_20c_2b_2b_20tpsa_20lib_56',['cppTPSA - C++ TPSA Lib',['../md__e___work__code_tpsa_lib_readme.html',1,'']]]
 ];
