+++ doc/doxygen/html/search/pages_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['c_2b_2b_20tpsa_20lib_204',['C++ TPSA Lib',['../md__e___work__code_tpsa_lib_readme.html',1,'']]]
+  ['cpptpsa_20_2d_20c_2b_2b_20tpsa_20lib_204',['cppTPSA - C++ TPSA Lib',['../md__e___work__code_tpsa_lib_readme.html',1,'']]]
 ];
