+++ doc/doxygen/latex/refman.tex
@@ -175,7 +175,7 @@
 \label{md__e___work__code_tpsa_lib__l_i_c_e_n_s_e}
 \Hypertarget{md__e___work__code_tpsa_lib__l_i_c_e_n_s_e}
 \input{md__e___work__code_tpsa_lib__l_i_c_e_n_s_e}
-\chapter{C++ TPSA Lib}
+\chapter{cpp\+TPSA -\/ C++ TPSA Lib}
 \label{md__e___work__code_tpsa_lib_readme}
 \Hypertarget{md__e___work__code_tpsa_lib_readme}
 \input{md__e___work__code_tpsa_lib_readme}

[…]
