### 2023-01-23 bc97aed
docs(component): fixed an error in readme on compiling the examples. -std=c++11 -> -std=c++14
