+++ readme.md
@@ -6,54 +6,44 @@
 
 This code allows users to do computations using Truncated Power Series Algebra (TPSA) and/or Differential Algebra (DA). A Python wrapper of this lib is hosted in [a separate repository](https://github.com/zhanghe9704/tpsa-python). 
 
-
-
 For TPSA and DA, please refer to chapter 8 in [*Lecture Notes on Special Topics in Accelerator Physics*](http://inspirehep.net/record/595287/files/slac-pub-9574.pdf)  by  Prof. Alex Chao  and chapter 2 in [*Modern Map Methods in Particle Beam Physics*](http://bt.pa.msu.edu/cgi-bin/display.pl?name=AIEP108book) by Prof. Martin Berz. 
 
-
-
 This code is developed based on Dr. Lingyun Yang's tpsa codes in C++ . His codes (tpsa.cpp and tpsa.h) are included in this repository. They are untouched, except for a few functions that are commented off and replaced by functions in tpsa_extend.cc. 
-
-
 
 The major change is the memory management. The current memory management works in the following way. Before any TPSA/DA calculation, one needs to reserve the memory for $n$ TPS vector. A memory pool is allocated in the heap, which can be considered as $n$ slots, each for one TPS vector. A linked list is created to reference the  available (unused) slots. Two pointers point to the beginning and the end of the linked list respectively. When a new TPS vector is created, the first slot will be assigned to it and the beginning pointer points to the  following slot in the linked list. When a TPS vector goes out of its scope, the slot will be reset to empty and liked to the end of the list. The end pointer will be updated and point to the new end.  
 
 ![Memory management](doc/tpsa_memory_management.png)
 
-
-
 A new data type DAVector is created as a wrapper of the TPS vector. The following mathematical operator and functions are overloaded for DAVector, so that a variable of DAVector type can be used as a intrinsic type in calculations. 
-
-
 
 Math operator overloaded: (DA - DA vector, CD - complex DA vector)
 
 | Left hand | Operator | Right hand |
-| :-------: | :------: | :--------: |
-|   DA/CD   |    +     |   DA/CD    |
-|  double   |    +     |   DA/CD    |
-|   DA/CD   |    +     |   double   |
-|           |    +     |   DA/CD    |
-|   DA/CD   |    -     |   DA/CD    |
-|   DA/CD   |    -     |   double   |
-|  double   |    -     |   DA/CD    |
-|           |    -     |   DA/CD    |
-|   DA/CD   |    *     |   DA/CD    |
-|   DA/CD   |    *     |   double   |
-|  double   |    *     |   DA/CD    |
-|   DA/CD   |    /     |   DA/CD    |
-|   DA/CD   |    /     |   double   |
-|  double   |    /     |   DA/CD    |
-|   DA/CD   |    =     |   DA/CD    |
-|   DA/CD   |    =     |   double   |
-|   DA/CD   |    +=    |   DA/CD    |
-|   DA/CD   |    +=    |   double   |
-|   DA/CD   |    -=    |   DA/CD    |
-|   DA/CD   |    -=    |   double   |
-|   DA/CD   |    *=    |   DA/CD    |
-|   DA/CD   |    *=    |   double   |
-|   DA/CD   |    /=    |   DA/CD    |
-|   DA/CD   |    /=    |   double   |
+|:---------:|:--------:|:----------:|
+| DA/CD     | +        | DA/CD      |
+| double    | +        | DA/CD      |
+| DA/CD     | +        | double     |
+|           | +        | DA/CD      |
+| DA/CD     | -        | DA/CD      |
[…]
