### 2023-01-23 b07df2a
docs(component): add a section about the Python wrapper with link to the repository in readme
