### 2023-01-23 24b16df
docs(component): add a paragraph on how to compile the tests
