### 2023-01-23 dca0fac
docs(component): document ADOrderTable and related functions, with a few other functions. Update doc
