+++ doc/doxygen/html/search/all_10.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['table_5fbegin_170',['table_begin',['../class_a_d_order_table.html#aa7a33812fa83049f3560b8d4c073abf0',1,'ADOrderTable']]],
-  ['table_5fend_171',['table_end',['../class_a_d_order_table.html#aa5fc114196f9aa2c83a3afa841e35421',1,'ADOrderTable']]],
-  ['tan_172',['tan',['../da_8h.html#a1cf3c2f3379994543b99ccda65fe9fc6',1,'tan(const DAVector &amp;da_vector):&#160;da.cc'],['../da_8cc.html#a1cf3c2f3379994543b99ccda65fe9fc6',1,'tan(const DAVector &amp;da_vector):&#160;da.cc']]],
-  ['tanh_173',['tanh',['../da_8h.html#af63923c4fbe9c11020fc8ea26a2f946b',1,'tanh(const DAVector &amp;da_vector):&#160;da.cc'],['../da_8cc.html#af63923c4fbe9c11020fc8ea26a2f946b',1,'tanh(const DAVector &amp;da_vector):&#160;da.cc']]],
-  ['tnvnd_174',['TNVND',['../tpsa_8h.html#a015d1e5f2a6fc8bd756badae78e77360',1,'tpsa.h']]],
-  ['tpsa_2ecpp_175',['tpsa.cpp',['../tpsa_8cpp.html',1,'']]],
-  ['tpsa_2eh_176',['tpsa.h',['../tpsa_8h.html',1,'']]],
-  ['tpsa_5fextend_2ecc_177',['tpsa_extend.cc',['../tpsa__extend_8cc.html',1,'']]],
-  ['tpsa_5fextend_2eh_178',['tpsa_extend.h',['../tpsa__extend_8h.html',1,'']]],
-  ['trim_5fwhitespace_179',['trim_whitespace',['../da_8h.html#a639c9fb588c015dc76640cf52662a3b5',1,'trim_whitespace(std::string input_line):&#160;da.cc'],['../da_8cc.html#a639c9fb588c015dc76640cf52662a3b5',1,'trim_whitespace(std::string input_line):&#160;da.cc']]],
-  ['tvec_180',['TVEC',['../tpsa_8h.html#a36fc79ba14099c25d212e5412c95806f',1,'tpsa.h']]]
+  ['table_5fbegin_178',['table_begin',['../class_a_d_order_table.html#aa7a33812fa83049f3560b8d4c073abf0',1,'ADOrderTable']]],
+  ['table_5fend_179',['table_end',['../class_a_d_order_table.html#aa5fc114196f9aa2c83a3afa841e35421',1,'ADOrderTable']]],
+  ['tan_180',['tan',['../da_8h.html#a1cf3c2f3379994543b99ccda65fe9fc6',1,'tan(const DAVector &amp;da_vector):&#160;da.cc'],['../da_8cc.html#a1cf3c2f3379994543b99ccda65fe9fc6',1,'tan(const DAVector &amp;da_vector):&#160;da.cc']]],
+  ['tanh_181',['tanh',['../da_8h.html#af63923c4fbe9c11020fc8ea26a2f946b',1,'tanh(const DAVector &amp;da_vector):&#160;da.cc'],['../da_8cc.html#af63923c4fbe9c11020fc8ea26a2f946b',1,'tanh(const DAVector &amp;da_vector):&#160;da.cc']]],
+  ['tnvnd_182',['TNVND',['../tpsa_8h.html#a015d1e5f2a6fc8bd756badae78e77360',1,'tpsa.h']]],
[…]
