### 2023-01-23 #4818 @zhanghe9704
Hello,

I want to say thanks to the reviewers and the editors and say sorry for
the delayed reply.
I just pushed an update to the repository. I think I have addressed all the
minor_blocker/blocker/major_blocker in the reviewer comments, except the
following:

1. MINOR_BLOCKER: the doxygen file seems to have the joss pdf mixed in with
the code documentation, which makes reading the documentation hard.
I am not sure what it means. Looks to me what it says does not happen in my
copy.

2. BLOCKER: includes should be done by properly specifying a directory in
the cmake file, rather than including relative paths.
I think the reviewer is telling me to treat the "includes" in a
proper/professional way in cmake file but I do not know what I should do.
Could anyone give me more instructions? I am actually quite new to cmake.

3. MINOR BLOCKER: The binary operators should be documented. For example,
how is truncation handled in a multiplication operation? How is addition of
two DAVectors having different order handled?

This is actually about the definition of the binary operators on the
TPSA/DA vectors. They are handled exactly as how they are defined in the
theory. Besides this one, the reviewer asked a few other questions related
to the TPSA/DA theory, e.g. how the base is defined. The code is targeted
to users who know the fundamental theory of TPSA/DA and  I understand it
could be confusing for people who do not know it. But I still think it is
outside the scope of the code documentation to explain the theory because
it is hard to explain in a few sentences. I mentioned the respective
chapters in two books in the readme file,  which are very good instructions
about the theory, and both are available for free on the authors' websites.
Adding a few lines of comments, I am afraid, will only confuse or mislead
people.

Thanks again to the reviewers and editors. Please let me know if there's
anything I overlooked or misunderstood.

He Zhang
Computational Accelerator Physicist
Center for Advanced Science of Accelerators
Jefferson Lab
(O) 757-269-6657
www.hezhang.us
