### 2023-04-30 #4818 @pibion
👋 @ashok-arora, would be interested in able to review this article for JOSS?  I realize the subject isn't your area of expertise, but another reviewer has indicated that the CMAKE file needs some work, and we're looking for a CMAKE expert to make sure the building follows best practices.
