### 2023-05-02 #4818 @pibion
@rmrsk, @benjaminbolling, @georgiastuart, would you be interested and able to review this JOSS submission?
