### 2023-05-08 #4818 @benjaminbolling
Cmake - I am not an expert at all, I have limited experience with it (building a few packages some years ago) and will hence during these two weeks also have to refresh my knowledge and understand best practices of it.
