### 2023-05-08 #4818 @pibion
@benjaminbolling no pressure on CMAKE, the build instructions tell you what to do and I don't want to put extra burden on reviewers!  I'll keep working on getting a CMAKE best-practices person on board for this review.
