### 2023-05-20 #4818 @benjaminbolling
Just an update: I have completed review of the other JOSS paper now, so I will start reviewing this paper and package now. I expect to have it finished by next weekend.
