### 2023-05-22 #4818 @pibion
@benjaminbolling you should be able to create your review checklist with `@editorialbot generate my checklist`
