### 2023-10-01 #4818 @arfon
@pibion @danielskatz – I've generated a list of most similar papers over here https://github.com/openjournals/joss-reviews/issues/3665#issuecomment-1741985624 that might be useful for identifying reviewers (suggest asking previous _authors_ first).
