### 2023-12-09 #4818 @danielskatz
@pibion is stepping down as a JOSS editor, so I'll take this over
