### 2024-01-08 #4818 @danielskatz
@ax3l, @abgandar, @ianhbell, @fwitte, @arosen93, @corentin-dev, @CFGrote, @mbkumar - would two of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Note that for a variety of reasons, we did start this review quite a while ago, but had many delays, so are now basically restarting from scratch and looking for new reviewers. I hope the initial review work that has been done, while not complete, has improved the software/paper and will lead this part being faster.
