### 2024-01-16 #4818 @zhanghe9704
Thanks for reviewing this submission.
I just downloaded the code in a zip file and compiled it successfully using
cmake in Ubuntu.
Would you please describe what you have done and what you have seen?
It could help me to figure out where the problem is.

Best,
He

On Tue, Jan 16, 2024 at 4:08 AM Carsten Fortmann-Grote <
***@***.***> wrote:

> I started my review over in https://github.com/zhanghe9704/tpsa. First
> issue is that installation does not proceed. This blocks further
> assessment, I'll wait until fixed.
>
