### 2024-01-17 #4818 @zhanghe9704
I found you've sent me a separate email about the error.
The code can be compiled successfully using gcc 9.4.
After upgrading to gcc 12.3, I was able to reproduce the error.
I'll look into it.

Thanks,
He

On Tue, Jan 16, 2024 at 4:08 AM Carsten Fortmann-Grote <
***@***.***> wrote:

> I started my review over in https://github.com/zhanghe9704/tpsa. First
> issue is that installation does not proceed. This blocks further
> assessment, I'll wait until fixed.
>
