### 2024-01-17 #4818 @CFGrote
great. imho it would be enough to put gcc<=9.4 or whichever other
(greater) compiler version works into the requirements. but do as you
please.

looking forward to try it out with gcc-9.4
