### 2024-01-17 #4818 @CFGrote
The minimal gcc I have on my system is 10.2.1. Just tested it and successfully built libtpsa.a
