### 2024-01-17 #4818 @CFGrote
@danielskatz , at the current state, I don't think this paper is compatible with JOSS criteria as it contains code that must not be redistributed (include/tpsa.h and src/tpsa.cpp) and is hence in contradiction with the repo's OS license (MIT). See here for related issue in the code's repo: https://github.com/zhanghe9704/tpsa/issues/7
