### 2024-01-17 #4818 @danielskatz
@CFGrote - thanks for finding this and pointing it out. I agree that this is an item that blocks JOSS publication.

@zhanghe9704 - Can I get your response on this, and how you would like to proceed?
