### 2024-01-17 #4818 @zhanghe9704
This goes back to about 10 years ago when Lingyun sent me his codes. He
allows me to put them on github.

I'll contact Lingyun and ask him whether we can put the codes under the OS
license (MIT).

Best,
He

On Wed, Jan 17, 2024 at 9:33 AM Daniel S. Katz ***@***.***>
wrote:

> @CFGrote <https://github.com/CFGrote> - thanks for finding this and
> pointing it out. I agree that this is an item that blocks JOSS publication.
>
> @zhanghe9704 <https://github.com/zhanghe9704> - Can I get your response
> on this, and how you would like to proceed?
>
