### 2024-01-17 #4818 @zhanghe9704
Lingyun agrees to put his code under OS license (MIT).
See the following email from him.

He

---------- Forwarded message ---------
From: Lingyun Yang ***@***.***>
Date: Wed, Jan 17, 2024 at 1:54 PM
Subject: Re: Question about your tpsa code license
To: Zhang He ***@***.***>

Hi He,

I am fine the the MIT OS licence as long as the Copyright line is
still in the source code:

Copyright(C) 2008 by Lingyun Yang

Email and website can be removed,
also adding the reference to my original paper would be greatly appreciated.

L. Yang, “Array Based Truncated Power Series Package”, Proc. ICAP'09,
San Francisco, CA, USA, Aug.-Sep. 2009,

-- Lingyun
