### 2024-01-23 #4818 @zhanghe9704
Updated the repo to fix the compiling error when using gcc 12.
The error is generated because gcc 12 does not know how to calculate the
exp() function for complex DA Vector.
(Somehow, the previous versions of gcc can handle it.)
I explicitly defined the exp() function for complex DA Vectors, using the
formula:
for z = x + i*y, exp(z) = exp(x)*(cos(y)+i*sin(y)).
Now the codes can be compiled with gcc 12.

Thanks,
He
