### 2024-01-31 #4818 @zhanghe9704
Hello,

I have updated my repo following   @mbkumar <https://github.com/mbkumar>'s
suggestions.
Now the two repos have been combined into one. pyTPSA can be installed
using PIP.
Readme files and the paper have been revised respectively.

Best,
He
