### 2024-02-13 #4818 @CFGrote
Side note, the pdf does not render in the github viewer, I get a
![image](https://github.com/openjournals/joss-reviews/assets/15476807/93f58290-a875-4d1b-83e2-e75e52f8975a)
