### 2024-02-13 #4818 @danielskatz
Thanks @CFGrote - I look forward to some fixes to the references from @zhanghe9704.

Also, fwiw, the pdf renders correctly in the github viewer for me.
