### 2024-02-15 #4818 @CFGrote
Citations look good now, I made a comment on reference formatting over in https://github.com/zhanghe9704/tpsa/issues/16
