### 2024-02-26 #4818 @danielskatz
@mbkumar - can you check off all the items on [your checklist](https://github.com/openjournals/joss-reviews/issues/4818#issuecomment-1894003353) now?
