### 2024-02-26 #4818 @danielskatz
@CFGrote - I see four items on [your checklist](https://github.com/openjournals/joss-reviews/issues/4818#issuecomment-1893214554) not checked off - can you let me know what is needed for you to be able to check them off?
