### 2024-02-28 #4818 @zhanghe9704
Merged the pull request. Thanks.
He

On Wed, Feb 28, 2024 at 8:39 AM Daniel S. Katz ***@***.***>
wrote:

> @zhanghe9704 <https://github.com/zhanghe9704> - I've proofread the paper,
> and have suggested changes in zhanghe9704/tpsa#17
> <https://github.com/zhanghe9704/tpsa/pull/17> - please merge this, or let
> me know what you disagree with, then we can proceed to the final steps.
>
