### 2024-02-29 #4818 @zhanghe9704
Version tag: v1.1.1
DOI: 10.5281/zenodo.10728770

Please let me know if anything is wrong or missing. Thanks a lot!
